temp: 宏管理 - 校验,正则修正,重构
